表示される問題と回答する問題の一致
