--- dementia_quiz_VR/Assets/C#/PostQuizTF.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 5ee3df38eada43fb8618e913cd76e373
-timeCreated: 1715588036
