+++ dementia_quiz_VR/Assets/Scenes/QuizScene.unity
@@ -359,7 +359,7 @@ GameObject:
   - component: {fileID: 537533445}
   - component: {fileID: 537533446}
   m_Layer: 5
-  m_Name: LeftSelect
+  m_Name: LeftSelecttext
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -847,7 +847,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 660832129}
-  - component: {fileID: 660832128}
+  - component: {fileID: 660832130}
   m_Layer: 0
   m_Name: QuizManager
   m_TagString: Untagged
@@ -855,21 +855,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &660832128
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 660832127}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9ca6cdd35ae7e4a2b9ad2c3a802f624f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  Quizname: {fileID: 1756069871}
-  Quizsel_1: {fileID: 537533446}
-  Quizsel_2: {fileID: 1508672914}
 --- !u!4 &660832129
 Transform:
   m_ObjectHideFlags: 0
@@ -885,6 +870,21 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &660832130
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 660832127}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9ca6cdd35ae7e4a2b9ad2c3a802f624f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Quizname: {fileID: 1756069871}
+  Quizsel_1: {fileID: 537533446}
+  Quizsel_2: {fileID: 1508672914}
 --- !u!1 &714019307
 GameObject:
   m_ObjectHideFlags: 0
@@ -899,7 +899,7 @@ GameObject:
   - component: {fileID: 714019310}
   - component: {fileID: 714019309}
   m_Layer: 5
-  m_Name: 1
+  m_Name: Button1
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -937,7 +937,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2a9cf0cf74e146b08b9e2c8004f7111d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  quiztf: {fileID: 1017028337}
+  quiztf: {fileID: 660832130}
   LorR: L
 --- !u!114 &714019310
 MonoBehaviour:
@@ -1177,50 +1177,6 @@ MonoBehaviour:
   m_CameraFloorOffsetObject: {fileID: 2074307190}
   m_RequestedTrackingOriginMode: 0
   m_CameraYOffset: 1.1176
---- !u!1 &1017028336
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1017028338}
-  - component: {fileID: 1017028337}
-  m_Layer: 0
-  m_Name: QuizTFManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1017028337
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1017028336}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5ee3df38eada43fb8618e913cd76e373, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!4 &1017028338
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1017028336}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1081148748
 GameObject:
   m_ObjectHideFlags: 0
@@ -2120,7 +2076,7 @@ GameObject:
   - component: {fileID: 1508672915}
   - component: {fileID: 1508672914}
   m_Layer: 5
-  m_Name: RightSelect
+  m_Name: RightSelecttext
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -2254,7 +2210,7 @@ GameObject:
   - component: {fileID: 1756069870}
   - component: {fileID: 1756069871}
   m_Layer: 5
-  m_Name: Text (TMP)
+  m_Name: Quiztext
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -2370,7 +2326,7 @@ MonoBehaviour:
   m_VertexBufferAutoSizeReduction: 0
   m_useMaxVisibleDescender: 1
   m_pageToDisplay: 1
-  m_margin: {x: 0, y: 0, z: 0, w: 0}
+  m_margin: {x: 0, y: 0, z: -69.445496, w: 0}
   m_isUsingLegacyAnimationComponent: 0
   m_isVolumetricText: 0
   m_hasFontAssetChanged: 0
@@ -2390,7 +2346,7 @@ GameObject:
   - component: {fileID: 1777510943}
   - component: {fileID: 1777510942}
   m_Layer: 5
-  m_Name: 2
+  m_Name: Button2
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -2428,7 +2384,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2a9cf0cf74e146b08b9e2c8004f7111d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  quiztf: {fileID: 1017028337}
+  quiztf: {fileID: 660832130}
   LorR: R
 --- !u!114 &1777510943
 MonoBehaviour:
@@ -3387,4 +3343,3 @@ SceneRoots:
   - {fileID: 2040135217}
   - {fileID: 96929504}
   - {fileID: 660832129}
-  - {fileID: 1017028338}
